Pull out Stop in its own file
